Use dataview base to get column type in formula dataviews
