+++ lib/cartodb/models/dataview/base.js
@@ -1,3 +1,6 @@
+var dot = require('dot');
+dot.templateSettings.strip = false;
+
 function BaseDataview() {}
 
 module.exports = BaseDataview;
@@ -24,3 +27,34 @@ BaseDataview.prototype.getResult = function(psql, override, callback) {
 BaseDataview.prototype.search = function(psql, userQuery, callback) {
     return callback(null, this.format({ rows: [] }));
 };
+
+var FLOAT_OIDS = {
+    700: true,
+    701: true
+};
+
+var columnTypeQueryTpl = dot.template(
+    'SELECT pg_typeof({{=it.column}})::oid FROM ({{=it.query}}) _cdb_column_type limit 1'
+);
+
+BaseDataview.prototype.getColumnType = function (psql, column, query, callback) {
+    var readOnlyTransaction = true;
+
+    var columnTypeQuery = columnTypeQueryTpl({
+        column: column, query: query
+    });
+
+    psql.query(columnTypeQuery, function(err, result) {
+        if (err) {
+            return callback(err);
+        }
+        var pgType = result.rows[0].pg_typeof;
+        callback(null, getPGTypeName(pgType));
+    }, readOnlyTransaction);
+};
+
+function getPGTypeName (pgType) {
+    return {
+        float: FLOAT_OIDS.hasOwnProperty(pgType)
+    };
+}
