+++ lib/cartodb/models/dataview/formula.js
@@ -5,10 +5,6 @@ var debug = require('debug')('windshaft:widget:formula');
 var dot = require('dot');
 dot.templateSettings.strip = false;
 
-var columnTypeQueryTpl = dot.template(
-    'SELECT pg_typeof({{=it.column}})::oid FROM ({{=it.query}}) _cdb_histogram_column_type limit 1'
-);
-
 var formulaQueryTpl = dot.template([
     'SELECT',
     '  {{=it._operation}}({{=it._column}}) AS result,',
@@ -32,11 +28,6 @@ var VALID_OPERATIONS = {
     sum: true,
     min: true,
     max: true
-};
-
-var FLOAT_OIDS = {
-    700: true,
-    701: true
 };
 
 var TYPE = 'formula';
@@ -84,30 +75,20 @@ Formula.prototype.sql = function(psql, override, callback) {
         override = {};
     }
 
-    var _query = this.query;
-
-    var columnTypeQuery = columnTypeQueryTpl({
-        column: this.column, query: _query
-    });
-
     if (this._isFloatColumn === null) {
-        var readOnlyTransaction = true;
-        psql.query(columnTypeQuery, function(err, result) {
-            self._isFloatColumn = false;
-            if (!err && !!result.rows[0]) {
-                var pgType = result.rows[0].pg_typeof;
-                if (FLOAT_OIDS.hasOwnProperty(pgType)) {
-                    self._isFloatColumn = true;
-                }
+        this._isFloatColumn = false;
+        this.getColumnType(psql, this.column, this.query, function (err, type) {
+            if (!err && !!type) {
+                self._isFloatColumn = type.float;
             }
             self.sql(psql, override, callback);
-        }, readOnlyTransaction);
+        });
         return null;
     }
 
     var formulaSql = formulaQueryTpl({
         _isFloatColumn: this._isFloatColumn,
-        _query: _query,
+        _query: this.query,
         _operation: this.operation,
         _column: this.column
     });
