+++ lib/cartodb/models/dataview/overviews/formula.js
@@ -5,58 +5,49 @@ var debug = require('debug')('windshaft:widget:formula:overview');
 var dot = require('dot');
 dot.templateSettings.strip = false;
 
-var FLOAT_OIDS = {
-    700: true,
-    701: true
-};
-
-var columnTypeQueryTpl = dot.template(
-    'SELECT pg_typeof({{=it.column}})::oid FROM ({{=it.query}}) _cdb_histogram_column_type limit 1'
-);
-
 var formulaQueryTpls = {
-  'count': dot.template([
-      'SELECT',
-      'sum(_feature_count) AS result,',
-      '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nulls WHERE {{=it._column}} IS NULL) AS nulls_count',
-      '{{?it._isFloatColumn}},(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_infinities',
-      '  WHERE {{=it._column}} = \'infinity\'::float OR {{=it._column}} = \'-infinity\'::float) AS infinities_count,',
-      '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nans',
-      '  WHERE {{=it._column}} = \'NaN\'::float) AS nans_count{{?}}',
-      'FROM ({{=it._query}}) _cdb_formula'
-  ].join('\n')),
-  'sum': dot.template([
-      'SELECT',
-      'sum({{=it._column}}*_feature_count) AS result,',
-      '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nulls WHERE {{=it._column}} IS NULL) AS nulls_count',
-      '{{?it._isFloatColumn}},(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_infinities',
-      '  WHERE {{=it._column}} = \'infinity\'::float OR {{=it._column}} = \'-infinity\'::float) AS infinities_count',
-      ',(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nans',
-      '  WHERE {{=it._column}} = \'NaN\'::float) AS nans_count{{?}}',
-      'FROM ({{=it._query}}) _cdb_formula',
-      '{{?it._isFloatColumn}}WHERE',
-      '  {{=it._column}} != \'infinity\'::float',
-      'AND',
-      '  {{=it._column}} != \'-infinity\'::float',
-      'AND',
-      '  {{=it._column}} != \'NaN\'::float{{?}}'
-  ].join('\n')),
-  'avg': dot.template([
-      'SELECT',
-      'sum({{=it._column}}*_feature_count)/sum(_feature_count) AS result,',
-      '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nulls WHERE {{=it._column}} IS NULL) AS nulls_count',
-      '{{?it._isFloatColumn}},(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_infinities',
-      '  WHERE {{=it._column}} = \'infinity\'::float OR {{=it._column}} = \'-infinity\'::float) AS infinities_count',
-      ',(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nans',
-      '  WHERE {{=it._column}} = \'NaN\'::float) AS nans_count{{?}}',
-      'FROM ({{=it._query}}) _cdb_formula',
-      '{{?it._isFloatColumn}}WHERE',
-      '  {{=it._column}} != \'infinity\'::float',
-      'AND',
-      '  {{=it._column}} != \'-infinity\'::float',
-      'AND',
-      '  {{=it._column}} != \'NaN\'::float{{?}}'
-  ].join('\n')),
+    'count': dot.template([
+        'SELECT',
+        'sum(_feature_count) AS result,',
+        '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nulls WHERE {{=it._column}} IS NULL) AS nulls_count',
+        '{{?it._isFloatColumn}},(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_infinities',
+        '  WHERE {{=it._column}} = \'infinity\'::float OR {{=it._column}} = \'-infinity\'::float) AS infinities_count,',
+        '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nans',
+        '  WHERE {{=it._column}} = \'NaN\'::float) AS nans_count{{?}}',
+        'FROM ({{=it._query}}) _cdb_formula'
+    ].join('\n')),
+    'sum': dot.template([
+        'SELECT',
+        'sum({{=it._column}}*_feature_count) AS result,',
+        '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nulls WHERE {{=it._column}} IS NULL) AS nulls_count',
+        '{{?it._isFloatColumn}},(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_infinities',
+        '  WHERE {{=it._column}} = \'infinity\'::float OR {{=it._column}} = \'-infinity\'::float) AS infinities_count',
+        ',(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nans',
+        '  WHERE {{=it._column}} = \'NaN\'::float) AS nans_count{{?}}',
+        'FROM ({{=it._query}}) _cdb_formula',
+        '{{?it._isFloatColumn}}WHERE',
+        '  {{=it._column}} != \'infinity\'::float',
+        'AND',
+        '  {{=it._column}} != \'-infinity\'::float',
+        'AND',
+        '  {{=it._column}} != \'NaN\'::float{{?}}'
+    ].join('\n')),
+    'avg': dot.template([
+        'SELECT',
+        'sum({{=it._column}}*_feature_count)/sum(_feature_count) AS result,',
+        '(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nulls WHERE {{=it._column}} IS NULL) AS nulls_count',
+        '{{?it._isFloatColumn}},(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_infinities',
+        '  WHERE {{=it._column}} = \'infinity\'::float OR {{=it._column}} = \'-infinity\'::float) AS infinities_count',
+        ',(SELECT count(1) FROM ({{=it._query}}) _cdb_formula_nans',
+        '  WHERE {{=it._column}} = \'NaN\'::float) AS nans_count{{?}}',
+        'FROM ({{=it._query}}) _cdb_formula',
+        '{{?it._isFloatColumn}}WHERE',
+        '  {{=it._column}} != \'infinity\'::float',
+        'AND',
+        '  {{=it._column}} != \'-infinity\'::float',
+        'AND',
+        '  {{=it._column}} != \'NaN\'::float{{?}}'
+    ].join('\n')),
 };
 
 function Formula(query, options, queryRewriter, queryRewriteData, params) {
@@ -71,38 +62,30 @@ Formula.prototype.constructor = Formula;
 
 module.exports = Formula;
 
-Formula.prototype.sql = function(psql, override, callback) {
+Formula.prototype.sql = function (psql, override, callback) {
     var self = this;
     var formulaQueryTpl = formulaQueryTpls[this.operation];
 
-    if ( formulaQueryTpl ) {
+    if (formulaQueryTpl) {
         // supported formula for use with overviews
-
-        var columnTypeQuery = columnTypeQueryTpl({
-            column: this.column, query: this.query
-        });
-
         if (this._isFloatColumn === null) {
-            var readOnlyTransaction = true;
-            psql.query(columnTypeQuery, function(err, result) {
-                self._isFloatColumn = false;
-                if (!err && !!result.rows[0]) {
-                    var pgType = result.rows[0].pg_typeof;
-                    if (FLOAT_OIDS.hasOwnProperty(pgType)) {
-                        self._isFloatColumn = true;
-                    }
+            this._isFloatColumn = false;
+            this.getColumnType(psql, this.column, this.query, function (err, type) {
+                if (!err && !!type) {
+                    self._isFloatColumn = type.float;
                 }
                 self.sql(psql, override, callback);
-            }, readOnlyTransaction);
+            });
             return null;
         }
 
         var formulaSql = formulaQueryTpl({
             _isFloatColumn: this._isFloatColumn,
-            _query:     this.rewrittenQuery(this.query),
+            _query: this.rewrittenQuery(this.query),
             _operation: this.operation,
-            _column:    this.column
+            _column: this.column
         });
+
         callback = callback || override;
 
         debug(formulaSql);
